Use enabled flags to disable tests instead
We used to naively exclude the FHIRPath test cases from the test suite. This makes it difficult to keep track of ignored test cases and the reason.

The enabled flag is more explicit and makes it easier to keep track of ignored tests: https://kotest.io/docs/framework/conditional/enabled-config-flag.html